Rename exp3.cpp to BubbleSortBinSearch.cpp
Write a program to accept N integers from the user and store them in an array. Sort the array in ascending order using Bubble sort. Then accept another number from the user, search whether that number exists in the array using Binary Search. If it does, display its index and if it doesn’t, then print that the number is not found in the array.